+++ app/src/main/java/com/example/myapplication1/core/model/contact/ContactModel.kt
@@ -9,28 +9,28 @@ data class Contact(val id:Int,val name:String, val phoneNumber:String, val profi
 object ContactModel {
     var contacts: List<Contact> =emptyList()
 
-    fun fetchContacts(context: Context): List<Contact> {
-        val contactList = mutableListOf<Contact>()
-
-        val cursor = context.contentResolver.query(
-            ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
-            null, null, null, null
-        )
-
-        cursor?.use {
-            val nameIndex = it.getColumnIndex(ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME)
-            val phoneNumberIndex = cursor.getColumnIndex(ContactsContract.CommonDataKinds.Phone.NUMBER)
-            val photoUriIndex = cursor.getColumnIndex(ContactsContract.CommonDataKinds.Phone.PHOTO_URI)
-            while (it.moveToNext()) {
-                val name = it.getString(nameIndex)
-                val phoneNumber = it.getString(phoneNumberIndex)
-                val photo = it.getString(photoUriIndex)?.toUri()
-                contactList.add(Contact(0, name, phoneNumber, photo?:Uri.EMPTY))
-            }
-        }
-
-        return contactList
-    }
+//    fun fetchContacts(context: Context): List<Contact> {
+//        val contactList = mutableListOf<Contact>()
+//
+//        val cursor = context.contentResolver.query(
+//            ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
+//            null, null, null, null
+//        )
+//
+//        cursor?.use {
+//            val nameIndex = it.getColumnIndex(ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME)
+//            val phoneNumberIndex = cursor.getColumnIndex(ContactsContract.CommonDataKinds.Phone.NUMBER)
+//            val photoUriIndex = cursor.getColumnIndex(ContactsContract.CommonDataKinds.Phone.PHOTO_URI)
+//            while (it.moveToNext()) {
+//                val name = it.getString(nameIndex)
+//                val phoneNumber = it.getString(phoneNumberIndex)
+//                val photo = it.getString(photoUriIndex)?.toUri()
+//                contactList.add(Contact(0, name, phoneNumber, photo?:Uri.EMPTY))
+//            }
+//        }
+//
+//        return contactList
+//    }
 }
